requirements.txt
aiohttp==3.8.1
python-dotenv==0.19.2
elevenlabs==0.2.11
nltk==3.6.7
pyttsx3==2.71
requests==2.26.0
speechrecognition==3.8.1
openai-whisper==20230117
pyaudio==0.2.11
flask[async]==3.1.0
noisereduce==3.0.3